requirements.txt
Django==4.2.3
mysql-connector-python==8.0.33
